binance_futures_monitor/requirements.txt
ccxt
ta
python-dotenv
loguru
pandas
numpy
aiohttp
jinja2
aiohttp_jinja2